Remove all old dependencies / tomcat 7 / embedded
Old build was actually recompiling code.  Since we no longer do that,
there is no reason to include all the dependencies that were required
for the build.

Add tomcat 7 update and now section out both runs of tomcat 7 and 8 so
it is easier to know which we are building for at any given time.

While we are at it, allow for embedded tomcat when needed.

And why not, increase mandatory version of maven to 3.1.1.