Fix enforcer plugin.  Maven appears to have a bug.
The OSS parent does not put the version for this plugin in plugin
management.  As a result, it overrides part of what is used.  This in
turn makes it appear as though we are not using the correct version and
we are not partially.  So this trade off warning ensures we have the
same version throughout.