[migrate] Remove the original build system
we no longer need to build from here.  This will be used for coordination of releases since builds all exist in releases area.  The readme will be reworked across the modules therefore is not complete at this time.  The tomcat 7, 85, and 9 modules contain very similar readme with config/parent just pointing to root of project at the comment.